requirements.txt
bcrypt==2.0.0
cffi==1.5.2
Flask-Cors==2.1.2
Flask-JWT==0.3.2
flask-mongoengine==0.7.5
Flask-WTF==0.12
mongoengine==0.10.6
pycparser==2.14
PyJWT==1.4.0
redis==2.10.5
wheel==0.24.0
WTForms==2.1
